+++ src/site/markdown/changelog.md
@@ -7,7 +7,9 @@ The changelog should give you a quick overview of what change between the differ
 ### New Features
 
 - [6](https://github.com/FHOOEAIST/imaging/issues/6) - Added support for ImageJ based images.
+- [7](https://github.com/FHOOEAIST/imaging/issues/7) - Added support for OpenIMAJ based images and added new method `ChannelType.isValidValue()`
 - [11](https://github.com/FHOOEAIST/imaging/issues/11) - New `ImageFactoryFactory`, that allows to create an image factory for a given type.
+- [12](https://github.com/FHOOEAIST/imaging/issues/12) - Added support for ND4J based images
 
 ### Dependencies
 - [13](https://github.com/FHOOEAIST/imaging/issues/13) - Update from seshat 1.0.1 to 1.1.0